change package and extract slack
