+++ watcher.go
@@ -1,4 +1,4 @@
-package lib
+package dnotifier
 
 import (
 	"log"
